requirements.txt
Django==5.1.2
PyYAML==6.0.2
django-model-utils==5.0.0
ipdb==0.13.13
psycopg2-binary==2.9.7